requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.1
pytz==2024.1

# AWS integration
boto3==1.34.68
botocore==1.34.68

# Data processing
pydantic==2.6.3  # For data validation and settings management

# Logging enhancements
structlog==24.1.0  # Structured logging

# Testing
pytest==7.4.4
pytest-mock==3.12.0
responses==0.24.1
pytest-cov==4.1.0  # For test coverage reporting

# Development tools
black==24.2.0  # Code formatting
isort==5.13.2  # Import sorting
flake8==7.0.0  # Linting
mypy==1.8.0  # Type checking